Link cmark statically.
libcmark.so is not symlinked properly in Arch Linux. It only has
libcmark.so -> libcmark.so.0.31.1 (for example). Anything that is
linked to libcmark will break after upgrading cmark. Usually it should
have a link libcmark.so.0 -> libcmark.so.0.31.1, so that if I upgrade
libcmark, I don’t have to recompile everything that is linked to it.